164 Add x-ray microservice
CL needs to be able to check PDFs for bad redactions.  Rather than
adding x-ray to CL we added a new microservice /utils/x-ray/pdf/ to
do so

Two new unit tests were added which sends PDFs to be checked.  The PDFs
to check were pulled from the x-ray repo.  The naming convention
requires 'yes' for PDFs with bad redactions and "no" when they do not.